# perf: cache mined commitment for quorum merkle root calculation