--- a/src/evo/smldiff.cpp
+++ b/src/evo/smldiff.cpp
@@ -54,12 +54,11 @@ bool CSimplifiedMNListDiff::BuildQuorumsDiff(const CBlockIndex* baseBlockIndex, 
     for (const auto& p : quorumHashes) {
         const auto& [llmqType, hash] = p;
         if (!baseQuorumHashes.count(p)) {
-            uint256 minedBlockHash;
-            llmq::CFinalCommitmentPtr qc = quorum_block_processor.GetMinedCommitment(llmqType, hash, minedBlockHash);
-            if (qc == nullptr) {
+            auto [qc, minedBlockHash] = quorum_block_processor.GetMinedCommitment(llmqType, hash);
+            if (minedBlockHash == uint256::ZERO) {
                 return false;
             }
-            newQuorums.emplace_back(*qc);
+            newQuorums.emplace_back(std::move(qc));
         }
     }