## Issue being fixed or feature implemented
Every time when list of quorum is changed, the quorum final commitments for multiple active quorums are re-read from database.
It's a heavy operation in a current implementation due to heavy database reading, heavy BLS deserialization and multiple in-memory copies.

## What was done?
 - add cache for active quorums, see `llmq::utils::InitQuorumsCache(qc_hashes_cached)`
 - use RVO for GetMinedCommitment, avoid extra allocation by unique_ptr
 - adjust return value of GetLastMinedCommitmentsPerQuorumIndexUntilBlock to avoid extra data copies and transformations


## How Has This Been Tested?
Time of calculation `CheckCbTxMerkleRoots` is shortened for 40%, and this step is a significant part of block validation.
Expected impact to total time for block validation and re-index is 3-4%.

`develop`:
```
[bench]             - CachedGetQcHashesQcIndexedHashes: 1.16ms [5.22s]
[bench]       - CheckCbTxMerkleRoots: 1.30ms [6.04s]
```
vs PR:

```
[bench]             - CachedGetQcHashesQcIndexedHashes: 1.20ms [2.65s]
[bench]       - CheckCbTxMerkleRoots: 1.34ms [3.44s]
```

## Breaking Changes
N/A


## Checklist:
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added or updated relevant unit/integration/functional/e2e tests
- [ ] I have made corresponding changes to the documentation
- [x] I have assigned this pull request to a milestone